Prevent multiple users connect the same social account
When user A registers and connects a social account, then updates his user email, they can register a new user account using the same social account.
The result is 2 users that has the same social account connected. 
When trying to login with that social account, it will login the first user (first connected account returned from the query)

By checking if there is no account found, and only then enter the registration block we can prevent that, and skip to login the correct user.